# feat(json): make Duration encoder allocation-free