Benchstat result:
```
name              old time/op    new time/op    delta
EncodeDuration-4     106ns ± 7%      69ns ± 2%   -35.36%  (p=0.000 n=15+15)

name              old alloc/op   new alloc/op   delta
EncodeDuration-4     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=15+15)

name              old allocs/op  new allocs/op  delta
EncodeDuration-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
```